pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-trading-research"
version = "0.1.0"
description = "A-share 13-point signal research system."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "akshare",
  "matplotlib",
  "numpy",
  "pandas",
  "pyarrow",
  "pyyaml",
  "yfinance",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
